refactor: decompose SettingsView into section sub-components
Extract 4 section components + LlmSettingsSection from the 1,246-line
SettingsView.tsx:

- GeneralSection: language, DB file, exchange rates, LLM settings
- CustomizationSection: theme, font size, sidebar toggles, tag colors
- FormatsSection: currency, number format, date format, first day of week
- AboutSection: version info, licenses, contributors, external links
- LlmSettingsSection: moved to its own file (was already a separate function)

Parent SettingsView keeps sidebar navigation, activeSection state,
global handlers (reset defaults, tooltip, openExternal), and renders
the appropriate section component.
