+++ app/src/features/settings/SettingsView.tsx
@@ -1,250 +1,34 @@
-import {
-  Settings,
-  Globe,
-  HelpCircle,
-  ExternalLink,
-  BookOpen,
-  Lightbulb,
-  Bug,
-} from "lucide-react";
+import { Settings } from "lucide-react";
 import "../../styles/Settings.css";
 import { useNumberFormat } from "../../stores/number-format";
 import { useTheme } from "../../stores/theme";
-import { formatNumberWithLocale } from "../../utils/format";
-import { CURRENCIES } from "../../utils/currencies";
-import CustomSelect from "../../components/ui/CustomSelect";
-import Switch from "../../components/ui/Switch";
 import ErrorBoundary from "../../components/layout/ErrorBoundary";
 import { useState, useEffect } from "react";
 import { rust } from "../../api/tauri-client";
 import { save } from "@tauri-apps/plugin-dialog";
 import { open } from "@tauri-apps/plugin-shell";
 import { useTranslation } from "react-i18next";
-import { AVAILABLE_LANGUAGES } from "../../i18n/i18n";
-import { formatDateForUI } from "../../utils/format";
-import { IS_RELEASE, APP_VERSION, APP_COMMIT } from "../../utils/version";
 
 import { useCustomRate } from "../../hooks/useCustomRate";
 import { useConfirm } from "../../stores/confirm";
 import { useToast } from "../../contexts/toast";
-import CONTRIBUTORS from "../../config/contributors";
-import THIRD_PARTY_LICENSES from "../../config/licenses";
-import {
-  ChevronDown,
-  ChevronUp,
-  Loader2,
-  Check,
-  AlertCircle,
-} from "lucide-react";
-import ExchangeRatesList from "../../components/shared/ExchangeRatesList";
 import useTagColors from "../../hooks/useTagColors";
-import {
-  TAG_COLOR_KEYS,
-  getColorClasses,
-  getColorDot,
-} from "../../config/tag-colors";
 import {
   APP_DEFAULTS,
   DEFAULT_SIDEBAR_VISIBILITY,
-  EXTERNAL_URLS,
   RESETTABLE_STORAGE_KEYS,
   STORAGE_KEYS,
 } from "../../constants/app";
 
-interface LlmSettings {
-  ollama_url?: string;
-  ollama_model?: string;
-}
-
-interface OllamaModel {
-  name: string;
-  size?: number;
-}
-
-interface LlmSettingsSectionProps {
-  showTooltip: (e: React.MouseEvent | React.FocusEvent) => void;
-  hideTooltip: (e: React.MouseEvent | React.FocusEvent) => void;
-}
+import GeneralSection from "./GeneralSection";
+import CustomizationSection from "./CustomizationSection";
+import FormatsSection from "./FormatsSection";
+import AboutSection from "./AboutSection";
 
 interface SettingsViewProps {
   activeSection?: "general" | "customization" | "formats" | "about";
   sidebarVisibility?: Record<string, boolean>;
   onChangeSidebarVisibility?: (visibility: Record<string, boolean>) => void;
-}
-
-function LlmSettingsSection({
-  showTooltip,
-  hideTooltip,
-}: LlmSettingsSectionProps) {
-  const { t } = useTranslation();
-  const [ollamaUrl, setOllamaUrl] = useState("http://localhost:11434");
-  const [ollamaModel, setOllamaModel] = useState("");
-  const [models, setModels] = useState<OllamaModel[]>([]);
-  const [connected, setConnected] = useState<boolean | null>(null);
-  const [loading, setLoading] = useState(false);
-  const confirm = useConfirm();
-  const { showToast } = useToast();
-
-  useEffect(() => {
-    rust.get_llm_settings().then((_s) => {
-      const s = _s as LlmSettings;
-      if (s.ollama_url) setOllamaUrl(s.ollama_url);
-      if (s.ollama_model) setOllamaModel(s.ollama_model);
-    });
-  }, []);
-
-  async function handleTestConnection() {
-    setLoading(true);
-    try {
-      await rust.set_llm_settings({ ollamaUrl, ollamaModel });
-      const ok = await rust.check_ollama_connection();
-      setConnected(ok as boolean);
-      if (ok) {
-        const list = (await rust.list_ollama_models()) as OllamaModel[];
-        setModels(list);
-      }
-    } catch {
-      setConnected(false);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function handleSaveModel(val: string | number) {
-    setOllamaModel(String(val));
-    try {
-      await rust.set_llm_settings({ ollamaUrl, ollamaModel: String(val) });
-    } catch {
-      // ignore
-    }
-  }
-
-  async function handleSaveUrl() {
-    try {
-      await rust.set_llm_settings({ ollamaUrl, ollamaModel });
-    } catch {
-      // ignore
-    }
-  }
-
-  async function handleClearHistory() {
-    const ok = await (
-      confirm as (
-        message: string,
-        options?: Record<string, unknown>,
-      ) => Promise<boolean>
-    )(t("chat.clear_history_confirm"), {
-      title: t("chat.clear_history"),
-      kind: "warning",
-    });
-    if (!ok) return;
-    try {
-      await rust.delete_all_conversations();
-      showToast(t("chat.clear_history"), { type: "success" });
-    } catch {
-      // ignore
-    }
-  }
-
-  return (
-    <>
-      <div className="flex items-center justify-between mt-6 mb-2">
-        <div className="label-with-help">
-          <span
-            className="help-wrapper"
-            data-tooltip={t("settings.tooltip.llm")}
-            role="button"
-            tabIndex={0}
-            aria-label={t("settings.tooltip.llm")}
-            onMouseEnter={showTooltip}
-            onFocus={showTooltip}
-            onMouseLeave={hideTooltip}
-            onBlur={hideTooltip}
-          >
-            <HelpCircle
-              className="w-4 h-4 text-slate-400 help-icon"
-              aria-hidden="true"
-            />
-          </span>
-          <label className="settings-label">{t("nav.ai_assistant")}</label>
-        </div>
-      </div>
-
-      <div className="space-y-3">
-        <div>
-          <label className="block text-xs font-medium text-slate-500 dark:text-slate-400 mb-1">
-            {t("chat.ollama_url")}
-          </label>
-          <div className="flex gap-2">
-            <input
-              type="text"
-              value={ollamaUrl}
-              onChange={(e) => setOllamaUrl(e.target.value)}
-              onBlur={handleSaveUrl}
-              className="flex-1 bg-white dark:bg-slate-700 text-slate-700 dark:text-white text-sm py-1 px-2 rounded border border-slate-200 dark:border-slate-700 hover:bg-slate-50 dark:hover:bg-slate-700 transition-all outline-none focus:ring-2 focus:ring-brand-500"
-            />
-            <button
-              onClick={handleTestConnection}
-              disabled={loading}
-              className="px-3 py-1 bg-slate-100 dark:bg-slate-700 text-slate-700 dark:text-slate-300 rounded text-xs hover:bg-slate-200 dark:hover:bg-slate-600 transition-colors disabled:opacity-50"
-            >
-              {loading ? (
-                <Loader2 className="w-3.5 h-3.5 animate-spin" />
-              ) : (
-                t("chat.test_connection")
-              )}
-            </button>
-          </div>
-          {connected !== null && (
-            <div className="mt-1 flex items-center gap-1.5">
-              {connected ? (
-                <>
-                  <Check className="w-3 h-3 text-emerald-500" />
-                  <span className="text-xs text-emerald-600 dark:text-emerald-400">
-                    {t("chat.connected")}
-                  </span>
-                </>
-              ) : (
-                <>
-                  <AlertCircle className="w-3 h-3 text-red-500" />
-                  <span className="text-xs text-red-600 dark:text-red-400">
-                    {t("chat.not_connected")}
-                  </span>
-                </>
-              )}
-            </div>
-          )}
-        </div>
-
-        {models.length > 0 && (
-          <div>
-            <label className="block text-xs font-medium text-slate-500 dark:text-slate-400 mb-1">
-              {t("chat.model")}
-            </label>
-            <div className="relative settings-select">
-              <CustomSelect
-                value={ollamaModel}
-                onChange={(v) => handleSaveModel(v)}
-                options={models.map((m) => ({
-                  value: m.name,
-                  label: m.name,
-                }))}
-                placeholder={t("chat.setup.select_model")}
-                fullWidth={false}
-              />
-            </div>
-          </div>
-        )}
-
-        <button
-          onClick={handleClearHistory}
-          className="text-xs text-red-600 dark:text-red-400 hover:underline"
-        >
-          {t("chat.clear_history")}
-        </button>
-      </div>
-    </>
-  );
 }
 
 export default function SettingsView({
@@ -265,14 +49,12 @@ export default function SettingsView({
     uiLanguage,
     setUiLanguage,
   } = useNumberFormat();
-  const { theme, setTheme } = useTheme();
+  const { setTheme } = useTheme();
   const [dbPath, setDbPath] = useState("");
   const { checkAndPrompt, dialog } = useCustomRate();
   const confirm = useConfirm();
   const { showToast } = useToast();
-  const [showAllLicenses, setShowAllLicenses] = useState(false);
-  const { tagColors, setTagColor, resetAll: resetTagColors } = useTagColors();
-  const [categories, setCategories] = useState<string[]>([]);
+  const { resetAll: resetTagColors } = useTagColors();
   const [fontSize, setFontSize] = useState(() => {
     try {
       const v = localStorage.getItem(STORAGE_KEYS.FONT_SIZE);
@@ -293,11 +75,6 @@ export default function SettingsView({
       console.error("Failed to apply font size:", e);
     }
   }, [fontSize]);
-
-  const example = formatNumberWithLocale(1234.56, locale, {
-    style: "currency",
-    currency: currency || APP_DEFAULTS.CURRENCY,
-  });
 
   useEffect(() => {
     let mounted = true;
@@ -399,33 +176,6 @@ export default function SettingsView({
     }
   }
 
-  useEffect(() => {
-    if (activeSection === "customization") {
-      (async () => {
-        try {
-          const cats = (await rust.get_categories()) as string[];
-          const all = cats.includes("Transfer") ? cats : ["Transfer", ...cats];
-          setCategories(all.sort((a: string, b: string) => a.localeCompare(b)));
-        } catch (e) {
-          console.error("Failed to fetch categories:", e);
-        }
-      })();
-    }
-  }, [activeSection]);
-
-  const _today = new Date();
-  const dateFormatOptions = [
-    { value: "YYYY-MM-DD", label: formatDateForUI(_today, "YYYY-MM-DD") },
-    { value: "YYYY/MM/DD", label: formatDateForUI(_today, "YYYY/MM/DD") },
-    { value: "MM/DD/YYYY", label: formatDateForUI(_today, "MM/DD/YYYY") },
-    { value: "DD/MM/YYYY", label: formatDateForUI(_today, "DD/MM/YYYY") },
-    { value: "DD-MM-YYYY", label: formatDateForUI(_today, "DD-MM-YYYY") },
-    { value: "DD.MM.YYYY", label: formatDateForUI(_today, "DD.MM.YYYY") },
-    { value: "DD MMM YYYY", label: formatDateForUI(_today, "DD MMM YYYY") },
-    { value: "MMM DD, YYYY", label: formatDateForUI(_today, "MMM DD, YYYY") },
-    { value: "MMMM D, YYYY", label: formatDateForUI(_today, "MMMM D, YYYY") },
-  ];
-
   return (
     <ErrorBoundary>
       <div className="settings-view">
@@ -448,783 +198,45 @@ export default function SettingsView({
           </div>
 
           {activeSection === "general" && (
-            <>
-              <div className="flex items-center justify-between mt-4">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.language")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.language")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.language")}
-                  </label>
-                </div>
-              </div>
-              <div className="relative settings-select">
-                <CustomSelect
-                  value={uiLanguage}
-                  onChange={(v) => setUiLanguage(String(v))}
-                  options={AVAILABLE_LANGUAGES.map(({ code, label }) => ({
-                    value: code,
-                    label,
-                  }))}
-                  placeholder={t("settings.select_language_placeholder")}
-                  fullWidth={false}
-                  data-testid="language-select"
-                />
-              </div>
-              <div className="flex items-center justify-between">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.database_file")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.database_file")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.database_file")}
-                  </label>
-                </div>
-              </div>
-              <div className="relative">
-                <div className="flex items-center gap-2">
-                  <button
-                    type="button"
-                    className="bg-white dark:bg-slate-700 text-slate-700 dark:text-white text-sm py-1 px-2 rounded w-full max-w-full text-left overflow-hidden truncate border border-slate-200 dark:border-slate-700 hover:bg-slate-50 dark:hover:bg-slate-700 transition-all"
-                    onClick={handleSelectDb}
-                    data-tooltip={dbPath || t("settings.select_db_file")}
-                    aria-label={dbPath || t("settings.select_db_file")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    {dbPath && dbPath.length > 0
-                      ? dbPath
-                      : t("settings.select_db_file")}
-                  </button>
-                </div>
-              </div>
-
-              <div className="flex items-center justify-between mt-6">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.exchange_rates")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.exchange_rates")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.exchange_rates")}
-                  </label>
-                </div>
-              </div>
-              <ExchangeRatesList />
-
-              <LlmSettingsSection
-                showTooltip={showTooltip}
-                hideTooltip={hideTooltip}
-              />
-            </>
+            <GeneralSection
+              uiLanguage={uiLanguage}
+              setUiLanguage={setUiLanguage}
+              dbPath={dbPath}
+              handleSelectDb={handleSelectDb}
+              showTooltip={showTooltip}
+              hideTooltip={hideTooltip}
+            />
           )}
 
-          {activeSection === "customization" && sidebarVisibility && (
-            <>
-              <div className="mt-6">
-                <div className="flex items-center justify-between">
-                  <div className="label-with-help">
-                    <span
-                      className="help-wrapper"
-                      data-tooltip={t("settings.tooltip.theme")}
-                      role="button"
-                      tabIndex={0}
-                      aria-label={t("settings.tooltip.theme")}
-                      onMouseEnter={showTooltip}
-                      onFocus={showTooltip}
-                      onMouseLeave={hideTooltip}
-                      onBlur={hideTooltip}
-                    >
-                      <HelpCircle
-                        className="w-4 h-4 text-slate-400 help-icon"
-                        aria-hidden="true"
-                      />
-                    </span>
-                    <label className="settings-label">
-                      {t("settings.theme")}
-                    </label>
-                  </div>
-                </div>
-
-                <div className="relative settings-select mt-2">
-                  <CustomSelect
-                    value={theme}
-                    onChange={(v) => setTheme(String(v))}
-                    options={[
-                      { value: "light", label: t("settings.theme.light") },
-                      {
-                        value: "high-contrast-light",
-                        label: t("settings.theme.high_contrast_light"),
-                      },
-                      { value: "dark", label: t("settings.theme.dark") },
-                      {
-                        value: "high-contrast-dark",
-                        label: t("settings.theme.high_contrast_dark"),
-                      },
-                      {
-                        value: "ink-light",
-                        label: t("settings.theme.ink_light"),
-                      },
-                      {
-                        value: "ink-dark",
-                        label: t("settings.theme.ink_dark"),
-                      },
-                      { value: "system", label: t("settings.theme.system") },
-                    ]}
-                    placeholder={t("settings.select_theme_placeholder")}
-                    fullWidth={false}
-                  />
-                </div>
-
-                <div className="flex items-center justify-between mt-4">
-                  <div className="label-with-help">
-                    <span
-                      className="help-wrapper"
-                      data-tooltip={t("settings.tooltip.font_size")}
-                      role="button"
-                      tabIndex={0}
-                      aria-label={t("settings.tooltip.font_size")}
-                      onMouseEnter={showTooltip}
-                      onFocus={showTooltip}
-                      onMouseLeave={hideTooltip}
-                      onBlur={hideTooltip}
-                    >
-                      <HelpCircle
-                        className="w-4 h-4 text-slate-400 help-icon"
-                        aria-hidden="true"
-                      />
-                    </span>
-                    <label className="settings-label">
-                      {t("settings.font_size")}
-                    </label>
-                  </div>
-                  <div className="text-sm text-slate-500">
-                    {Math.round(fontSize * 100)}%
-                  </div>
-                </div>
-                <div className="relative mt-1 settings-slider">
-                  <input
-                    type="range"
-                    min={0.75}
-                    max={1.25}
-                    step={0.05}
-                    value={fontSize}
-                    onChange={(e) => setFontSize(Number(e.target.value))}
-                    className="w-full accent-brand-500"
-                    aria-label={t("settings.font_size")}
-                  />
-                </div>
-              </div>
-
-              <div className="flex items-center justify-between mb-4">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.sidebar_items")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.sidebar_items")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.sidebar_items")}
-                  </label>
-                </div>
-              </div>
-
-              <div className="bg-white dark:bg-slate-800 rounded-lg border border-slate-200 dark:border-slate-700 divide-y divide-slate-100 dark:divide-slate-700 overflow-hidden">
-                {[
-                  {
-                    key: "dashboard",
-                    label: t("settings.sidebar.dashboard"),
-                  },
-                  {
-                    key: "investments",
-                    label: t("settings.sidebar.investments"),
-                  },
-                  {
-                    key: "fire",
-                    label: t("settings.sidebar.fire_calculator"),
-                  },
-                  { key: "rules", label: t("settings.sidebar.rules") },
-                  {
-                    key: "scheduled",
-                    label: t("settings.sidebar.scheduled"),
-                  },
-                  {
-                    key: "chat",
-                    label: t("settings.sidebar.ai_assistant"),
-                  },
-                  {
-                    key: "all",
-                    label: t("settings.sidebar.all_transactions"),
-                  },
-                ].map(({ key, label }) => (
-                  <div
-                    key={key}
-                    className="flex items-center justify-between p-3 hover:bg-slate-50 dark:hover:bg-slate-700/50 transition-colors"
-                  >
-                    <span className="text-sm font-medium text-slate-700 dark:text-slate-300">
-                      {label}
-                    </span>
-                    <Switch
-                      checked={sidebarVisibility[key]}
-                      onChange={(val) =>
-                        onChangeSidebarVisibility?.({
-                          ...sidebarVisibility,
-                          [key]: val,
-                        })
-                      }
-                      aria-label={label}
-                    />
-                  </div>
-                ))}
-              </div>
-
-              <div className="flex items-center justify-between mt-6 mb-4">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.tag_colors")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.tag_colors")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.tag_colors")}
-                  </label>
-                </div>
-              </div>
-
-              {categories.length === 0 ? (
-                <p className="text-sm text-slate-400 dark:text-slate-500 italic">
-                  {t("settings.tag_colors.empty")}
-                </p>
-              ) : (
-                <div className="bg-white dark:bg-slate-800 rounded-lg border border-slate-200 dark:border-slate-700 divide-y divide-slate-100 dark:divide-slate-700 overflow-hidden">
-                  {categories.map((cat) => (
-                    <div
-                      key={cat}
-                      className="flex items-center justify-between p-3 hover:bg-slate-50 dark:hover:bg-slate-700/50 transition-colors"
-                    >
-                      <span
-                        className={`px-2 py-1 inline-flex text-xs font-bold rounded-lg border ${
-                          tagColors[cat]
-                            ? getColorClasses(tagColors[cat])
-                            : cat === "Transfer"
-                              ? getColorClasses("purple")
-                              : getColorClasses("slate")
-                        }`}
-                      >
-                        {cat}
-                      </span>
-                      <div className="flex items-center gap-1">
-                        {TAG_COLOR_KEYS.map((colorKey) => (
-                          <button
-                            key={colorKey}
-                            type="button"
-                            onClick={() => setTagColor(cat, colorKey)}
-                            title={colorKey}
-                            className={`w-5 h-5 rounded-full border-2 transition-transform ${getColorDot(colorKey)} ${
-                              tagColors[cat] === colorKey ||
-                              (!tagColors[cat] &&
-                                ((cat === "Transfer" &&
-                                  colorKey === "purple") ||
-                                  (cat !== "Transfer" && colorKey === "slate")))
-                                ? "border-slate-900 dark:border-white scale-110"
-                                : "border-transparent hover:scale-110"
-                            }`}
-                            aria-label={colorKey}
-                          />
-                        ))}
-                        {/* cross reset button removed per user request; default color can still be restored by clearing storage or via reset-all setting */}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </>
+          {activeSection === "customization" && sidebarVisibility && onChangeSidebarVisibility && (
+            <CustomizationSection
+              sidebarVisibility={sidebarVisibility}
+              onChangeSidebarVisibility={onChangeSidebarVisibility}
+              showTooltip={showTooltip}
+              hideTooltip={hideTooltip}
+              fontSize={fontSize}
+              setFontSize={setFontSize}
+            />
           )}
 
           {activeSection === "formats" && (
-            <>
-              <div className="flex items-center justify-between">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.currency")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.currency")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.currency")}
-                  </label>
-                </div>
-              </div>
-              <div className="relative settings-select">
-                <CustomSelect
-                  value={currency}
-                  onChange={async (v) => {
-                    setCurrency(String(v));
-                    if (v) {
-                      const confirmed = await checkAndPrompt(String(v));
-                      if (!confirmed) {
-                        setCurrency(currency);
-                      }
-                    }
-                  }}
-                  options={CURRENCIES.map((c) => ({
-                    value: c.code,
-                    label: `${c.code} - ${c.name} (${c.symbol})`,
-                  }))}
-                  placeholder={t("account.placeholder.select_currency")}
-                  fullWidth={false}
-                />
-              </div>
-
-              <div className="flex items-center justify-between">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.number_format")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.number_format")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.number_format")}
-                  </label>
-                </div>
-              </div>
-
-              <div className="relative settings-select">
-                <CustomSelect
-                  value={locale}
-                  onChange={(v) => setLocale(String(v))}
-                  options={[
-                    { value: "en-US", label: "1,234.56" },
-                    { value: "de-DE", label: "1.234,56" },
-                    { value: "fr-FR", label: "1 234,56" },
-                    { value: "de-CH", label: "1'234.56" },
-                    { value: "en-IN", label: "1,23,456.78" },
-                  ]}
-                  placeholder={t("settings.select_format_placeholder")}
-                  fullWidth={false}
-                />
-              </div>
-              <p className="text-slate-400 mt-3">
-                {t("settings.example", { example })}
-              </p>
-
-              <div className="flex items-center justify-between mt-4">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.date_format")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.date_format")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.date_format")}
-                  </label>
-                </div>
-              </div>
-              <div className="relative settings-select">
-                <CustomSelect
-                  value={dateFormat}
-                  onChange={(v) => setDateFormat(String(v))}
-                  options={dateFormatOptions}
-                  placeholder={t("settings.select_date_format_placeholder")}
-                  fullWidth={false}
-                />
-              </div>
-
-              <div className="flex items-center justify-between mt-4">
-                <div className="label-with-help">
-                  <span
-                    className="help-wrapper"
-                    data-tooltip={t("settings.tooltip.first_day_of_week")}
-                    role="button"
-                    tabIndex={0}
-                    aria-label={t("settings.tooltip.first_day_of_week")}
-                    onMouseEnter={showTooltip}
-                    onFocus={showTooltip}
-                    onMouseLeave={hideTooltip}
-                    onBlur={hideTooltip}
-                  >
-                    <HelpCircle
-                      className="w-4 h-4 text-slate-400 help-icon"
-                      aria-hidden="true"
-                    />
-                  </span>
-                  <label className="settings-label">
-                    {t("settings.first_day_of_week")}
-                  </label>
-                </div>
-              </div>
-              <div className="relative settings-select">
-                <CustomSelect
-                  value={firstDayOfWeek}
-                  onChange={(v) => setFirstDayOfWeek(Number(v))}
-                  options={[
-                    { value: 1, label: t("weekday.monday") },
-                    { value: 2, label: t("weekday.tuesday") },
-                    { value: 3, label: t("weekday.wednesday") },
-                    { value: 4, label: t("weekday.thursday") },
-                    { value: 5, label: t("weekday.friday") },
-                    { value: 6, label: t("weekday.saturday") },
-                    { value: 0, label: t("weekday.sunday") },
-                  ]}
-                  placeholder={t("settings.select_first_day_placeholder")}
-                  fullWidth={false}
-                />
-              </div>
-            </>
+            <FormatsSection
+              locale={locale}
+              setLocale={setLocale}
+              currency={currency}
+              setCurrency={setCurrency}
+              dateFormat={dateFormat}
+              setDateFormat={setDateFormat}
+              firstDayOfWeek={firstDayOfWeek}
+              setFirstDayOfWeek={setFirstDayOfWeek}
+              checkAndPrompt={checkAndPrompt}
+              showTooltip={showTooltip}
+              hideTooltip={hideTooltip}
+            />
           )}
 
           {activeSection === "about" && (
-            <>
-              <div className="about-header">
-                <img
-                  src="/icon.png"
-                  alt="HoneyBear Folio"
-                  className="w-16 h-16 object-contain mb-3"
-                />
-                <h3 className="about-app-name">HoneyBear Folio</h3>
-                <div className="about-version-badge">
-                  <span>{t("about.version")}:</span>
-                  {IS_RELEASE && APP_VERSION ? (
-                    <>
-                      <a
-                        href={`${EXTERNAL_URLS.GITHUB_REPO}/releases/tag/v${APP_VERSION}`}
-                        className="about-version-link"
-                        onClick={(e) => {
-                          e.preventDefault();
-                          openExternal(
-                            `${EXTERNAL_URLS.GITHUB_REPO}/releases/tag/v${APP_VERSION}`,
-                          );
-                        }}
-                      >
-                        v{APP_VERSION}
-                      </a>
-                      {APP_COMMIT && (
-                        <>
-                          <p>
-                            (
-                            <a
-                              href={`${EXTERNAL_URLS.GITHUB_REPO}/commit/${APP_COMMIT}`}
-                              className="about-version-link"
-                              style={{ fontFamily: "monospace" }}
-                              onClick={(e) => {
-                                e.preventDefault();
-                                openExternal(
-                                  `${EXTERNAL_URLS.GITHUB_REPO}/commit/${APP_COMMIT}`,
-                                );
-                              }}
-                            >
-                              {APP_COMMIT.substring(0, 7)}
-                            </a>
-                            )
-                          </p>
-                        </>
-                      )}
-                    </>
-                  ) : (
-                    <>
-                      <span>dev</span>
-                      {APP_COMMIT && (
-                        <p>
-                          (
-                          <a
-                            href={`${EXTERNAL_URLS.GITHUB_REPO}/commit/${APP_COMMIT}`}
-                            className="about-version-link"
-                            style={{ fontFamily: "monospace" }}
-                            onClick={(e) => {
-                              e.preventDefault();
-                              openExternal(
-                                `${EXTERNAL_URLS.GITHUB_REPO}/commit/${APP_COMMIT}`,
-                              );
-                            }}
-                          >
-                            {APP_COMMIT.substring(0, 7)}
-                          </a>
-                          )
-                        </p>
-                      )}
-                    </>
-                  )}
-                </div>
-              </div>
-
-              <div className="about-section">
-                <h4 className="about-section-title">{t("about.copyright")}</h4>
-                <p className="about-section-content">© 2026 HoneyBearFolio</p>
-              </div>
-
-              <div className="about-section">
-                <h4 className="about-section-title">{t("about.license")}</h4>
-                <p className="about-license-text">{t("about.license_text")}</p>
-                <a
-                  href={EXTERNAL_URLS.LICENSE}
-                  className="about-link"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    openExternal(EXTERNAL_URLS.LICENSE);
-                  }}
-                >
-                  <ExternalLink className="w-3.5 h-3.5" />
-                  {t("about.view_license")}
-                </a>
-              </div>
-
-              <div className="about-section">
-                <h4 className="about-section-title">
-                  {t("about.third_party")}
-                </h4>
-                {showAllLicenses && (
-                  <ul className="about-license-list">
-                    {THIRD_PARTY_LICENSES.map((l) => (
-                      <li key={l.name}>
-                        <a
-                          href={l.url}
-                          className="about-link"
-                          onClick={(e) => {
-                            e.preventDefault();
-                            openExternal(l.url);
-                          }}
-                        >
-                          {l.name}
-                        </a>
-                        <span className="about-license-meta">
-                          ({l.license})
-                        </span>
-                      </li>
-                    ))}
-                  </ul>
-                )}
-
-                <div className="mt-2">
-                  <button
-                    onClick={() => setShowAllLicenses(!showAllLicenses)}
-                    className="inline-flex items-center gap-1 text-xs px-2 py-1 rounded-full bg-slate-200 dark:bg-slate-600 text-slate-600 dark:text-slate-200 hover:bg-slate-300 dark:hover:bg-slate-500 transition-colors"
-                  >
-                    {showAllLicenses ? (
-                      <>
-                        <span>{t("about.third_party_hide")}</span>
-                        <ChevronUp className="w-3 h-3" />
-                      </>
-                    ) : (
-                      <>
-                        <span>
-                          {t("about.third_party_show", {
-                            count: THIRD_PARTY_LICENSES.length,
-                          })}
-                        </span>
-                        <ChevronDown className="w-3 h-3" />
-                      </>
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              <div className="about-divider" />
-
-              <div className="about-section">
-                <h4 className="about-section-title">
-                  {t("about.contributors")}
-                </h4>
-                {CONTRIBUTORS.map((c) => {
-                  const profileUrl =
-                    c.github || `https://github.com/${c.username}`;
-                  const avatarUrl = `https://avatars.githubusercontent.com/${c.username}?s=120&v=4`;
-                  return (
-                    <a
-                      key={c.username}
-                      href={profileUrl}
-                      className="about-contributor about-contributor-link"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        openExternal(profileUrl);
-                      }}
-                    >
-                      <img
-                        src={avatarUrl}
-                        alt={`${c.username} avatar`}
-                        className="about-contributor-avatar"
-                      />
-                      <div className="about-contributor-info">
-                        <span className="about-contributor-name">
-                          {c.username}
-                        </span>
-                        <span className="about-contributor-role">
-                          {t(c.roleKey)}
-                        </span>
-                      </div>
-                    </a>
-                  );
-                })}
-              </div>
-
-              <div className="about-divider" />
-
-              <div className="about-section">
-                <div className="about-links">
-                  <a
-                    href={EXTERNAL_URLS.WEBSITE}
-                    className="about-link"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      openExternal(EXTERNAL_URLS.WEBSITE);
-                    }}
-                  >
-                    <Globe className="w-3.5 h-3.5" />
-                    {t("about.website")}
-                  </a>
-                  <a
-                    href={EXTERNAL_URLS.GITHUB_REPO}
-                    className="about-link"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      openExternal(EXTERNAL_URLS.GITHUB_REPO);
-                    }}
-                  >
-                    <svg
-                      className="w-3.5 h-3.5"
-                      viewBox="0 0 24 24"
-                      fill="currentColor"
-                      aria-hidden="true"
-                    >
-                      <path d="M12 0C5.37 0 0 5.37 0 12c0 5.3 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61-.546-1.385-1.335-1.755-1.335-1.755-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C20.565 21.795 24 17.295 24 12c0-6.63-5.37-12-12-12z" />
-                    </svg>
-                    {t("about.github")}
-                  </a>
-                  <a
-                    href={`${EXTERNAL_URLS.GITHUB_REPO}/issues/new?template=feature_request.md`}
-                    className="about-link"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      openExternal(
-                        `${EXTERNAL_URLS.GITHUB_REPO}/issues/new?template=feature_request.md`,
-                      );
-                    }}
-                  >
-                    <Lightbulb className="w-3.5 h-3.5" />
-                    {t("about.features")}
-                  </a>
-                  <a
-                    href={`${EXTERNAL_URLS.GITHUB_REPO}/issues/new?template=bug_report.md`}
-                    className="about-link"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      openExternal(
-                        `${EXTERNAL_URLS.GITHUB_REPO}/issues/new?template=bug_report.md`,
-                      );
-                    }}
-                  >
-                    <Bug className="w-3.5 h-3.5" />
-                    {t("about.issues")}
-                  </a>
-                  <a
-                    href={`${EXTERNAL_URLS.DOCS}`}
-                    className="about-link"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      openExternal(`${EXTERNAL_URLS.DOCS}`);
-                    }}
-                  >
-                    <BookOpen className="w-3.5 h-3.5" />
-                    {t("about.docs")}
-                  </a>
-                </div>
-              </div>
-            </>
+            <AboutSection openExternal={openExternal} />
           )}
         </div>
